==> analysis/10_numtraits_figure.R <==
# Generate Figure 2: degree of pleiotropy and relationship with variant annotations
rm(list=ls())
library(dplyr)
library(data.table)
library(ggplot2)
library(ggthemes)
library(gridExtra)

load("snps_clumped_combn.rda")

# Load chromatin states of 2,293 lead SNPs generated by the IDEAS method (Zhang, Yu, et al. "Jointly characterizing epigenetic dynamics across multiple human cell types." Nucleic acids research 44.14 (2016): 6721-6731.)
ideas.bed <- fread("functional/ideas_loci_snps_all.bed")
colnames(ideas.bed) <- c("CHR","BP","BPplus1","SNP","pval","pheno","numtraits","CHR_ideas","BP_start_ideas","BP_end_ideas","state", "eid")

ideas.active <- ideas.bed %>% filter(state%in%c("10_TssA","8_TssAFlnk","2_TxWk","14_TssWk","4_Enh","6_EnhG","17_EnhGA","5_Tx")) %>%
    group_by(SNP) %>%
    summarize(num_eid = length(unique(eid)))
snps_clumped_combn$numcelltype <- 0
snps_clumped_combn$numcelltype[match(ideas.active$SNP,snps_clumped_combn$SNP)] <- 
    ideas.active$num_eid

ideas.bed <- ideas.bed %>% filter(state%in%c("10_TssA","8_TssAFlnk","2_TxWk","14_TssWk","4_Enh","6_EnhG","17_EnhGA","5_Tx")) %>%
    mutate(state=ifelse(state=="10_TssA"|state=="8_TssAFlnk"|state=="14_TssWk", 
                        "10_TssA/8_TssAFlnk/\n14_TssWk", state),
           state=ifelse(state=="5_Tx"|state=="2_TxWk", "5_Tx/2_TxWk", state),
           state=ifelse(state=="6_EnhG"|state=="17_EnhGA", "6_EnhG/17_EnhGA", state))
df <-  ideas.bed %>% group_by(SNP, state) %>%
    summarize(num_eid = n())
df$numtraits_grp <- snps_clumped_combn$numtraits_grp[match(df$SNP,snps_clumped_combn$SNP)]
df$state <- factor(df$state, levels=c("10_TssA/8_TssAFlnk/\n14_TssWk","5_Tx/2_TxWk","4_Enh","6_EnhG/17_EnhGA"))


xlabel <- "Number of traits in GRASP + consortia"
margin.size <- 10

gg1 <- ggplot(data = snps_clumped_combn, aes(x=numtraits)) +
    geom_histogram(breaks=seq(0.5,max(snps_clumped_combn$numtraits),by=1)) +
    theme_classic() +
    labs(x=xlabel, y="Count") +
    theme(plot.margin=margin(margin.size,margin.size,margin.size,margin.size),
          text = element_text(size=15),
          axis.title.x = element_text(size=12))

gg2 <- ggplot(data = snps_clumped_combn, aes(x=numtraits/116,y=numtraits.ukb/4114)) +
    geom_point() +
    geom_smooth(method='lm', formula= y~x) +
    theme_classic() +
    geom_text(x=0.09, y=0.075, size = 4,
              label=paste0("Spearman correlation = ",round(with(snps_clumped_combn,cor(numtraits/116,numtraits.ukb/4114,use="complete.obs",method="spearman")),2))) +
    labs(x="% of traits in GRASP + consortia", y="% of traits in UKB") +
    theme(plot.margin=margin(margin.size,margin.size,margin.size,margin.size),
          text = element_text(size=15),
          axis.title.x = element_text(size=12))

gg3 <- ggplot(data = snps_clumped_combn, aes(x=numtraits,y=ldscore)) +
    geom_point() +
    geom_smooth(method='lm', formula= y~x) +
    theme_classic() +
    geom_text(x=5, y=2900, size = 4,
              label=paste0("R2 = ",round(with(snps_clumped_combn,cor(numtraits,ldscore,use="complete.obs")^2),3))) +
    labs(x=xlabel, y="LD score") +
    theme(plot.margin=margin(margin.size,margin.size,margin.size,margin.size),
          text = element_text(size=15),
          axis.title.x = element_text(size=12))

df1 <- snps_clumped_combn %>% select(numtraits_grp, num.eq = numtissues) %>%
    mutate(yname = "Tissues")
df2 <- snps_clumped_combn %>% select(numtraits_grp, num.eq = numegenes) %>%
    mutate(yname = "eGenes")
df.eq <- rbind(df1,df2)
gg4 <- ggplot(data = df.eq, aes(x=numtraits_grp,y=num.eq, fill=yname)) +
    geom_boxplot() +
    theme_classic() +
    theme(legend.title = element_blank(), plot.margin=margin(margin.size,margin.size,margin.size,margin.size),
          text = element_text(size=15)) +
    labs(x=xlabel, y="Num. of eGenes/tissues in GTEx")

df.joint <- snps_clumped_combn %>% 
    select(numtraits,ldscore,numtissues,numegenes,numcelltype) %>%
    scale()
mod <- summary(lm(numtraits~ldscore+numtissues+numegenes+numcelltype, data=data.frame(df.joint)))$coefficients
df.coef <- data.frame(annot = rownames(mod), mod, stringsAsFactors = FALSE)[-1,] %>%
    rename(est = Estimate, se = Std..Error) %>% 
    mutate(lower = est-1.96*se, upper = est+1.96*se)
df.coef$annot <- c("LD score","Num. of tissues\n as eQTL",
                   "Num. of eGenes","Num. of cell types\n in active chromatin")

gg5 <- ggplot(data=df.coef, aes(x=annot, y=est, ymin=lower, ymax=upper)) +
    geom_pointrange() + 
    geom_hline(yintercept=0, lty=2) +  # add a dotted line at x=1 after flip
    coord_flip() +  # flip coordinates (puts labels on y axis)
    labs(x="Annotation", y="Effect on pleiotropy (95% CI)") +
    theme_classic()  # use a white background


gg6 <- ggplot(data = df, aes(x=numtraits_grp,y=num_eid, fill=state)) +
    geom_boxplot() +
    theme_classic() +
    theme(legend.title = element_blank(), plot.margin=margin(margin.size,margin.size,margin.size,margin.size),
          text = element_text(size=15), strip.background = element_blank(),
          axis.text.x = element_text(angle=-45,hjust=0)) +
    facet_wrap(~state, nrow=1) +
    labs(x=xlabel, y="Num. of cell types in Roadmap")


png(filename = "numtraits_relationship_w_xlabel.png", 
    width=3400, height = 3400, res=300, type = "cairo")
grid.arrange(gg1,gg2,gg3,
             gg4, gg5,
             gg6, 
             ncol = 3, nrow = 3, 
             layout_matrix = rbind(c(1,2,3),c(4,4,5),c(6,6,6)))
dev.off()



## Based on ChromHMM
load("functional/df_allactive_chromstate.rda")
gg5 <- ggplot(data = df, aes(x=numtraits_grp,y=num_eid, fill=state)) +
    geom_boxplot() +
    theme_classic() +
    theme(legend.title = element_blank(), plot.margin=margin(margin.size,margin.size,margin.size,margin.size),
          text = element_text(size=15), strip.background = element_blank(),
          axis.text.x = element_text(angle=-45,hjust=0)) +
    facet_wrap(~state, nrow=2) +
    labs(x=xlabel, y="Num. of cell types in Roadmap")
png(filename = "numtraits_ChromHMM.png", 
    width=2800, height = 2000, res=300, type = "cairo")
gg5
dev.off()

## Output list of trait-specific (1 trait) and highly pleiotropic SNPS (>15 traits)
snps_clumped_combn %>% filter(numtraits==1) %>% 
    mutate(traits.ukb.name =  lapply(traits.ukb.name, function(x) gsub(","," -",x))) %>%
    mutate(traits.ukb.name = sapply(traits.ukb.name, paste, collapse=", "),
           traits.ukb = sapply(traits.ukb, paste, collapse=", ")) %>%
    fwrite("loci_1trait.csv", row.names=FALSE)

snps_clumped_combn %>% filter(numtraits>15) %>% 
    mutate(traits.ukb.name =  lapply(traits.ukb.name, function(x) gsub(","," -",x))) %>%
    mutate(traits.ukb.name = sapply(traits.ukb.name, paste, collapse=", "),
           traits.ukb = sapply(traits.ukb, paste, collapse=", "),
           pheno = sapply(pheno, paste, collapse=", ")) %>%
    fwrite("loci_morethan15traits.csv", row.names=FALSE)
